Add DeprecatedDeploy to allow global resources with a warning
