+++ exe/kubernetes-deploy
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require 'krane/deploy_task'
+require 'krane/deprecated_deploy_task'
 require 'krane/options_helper'
 require 'krane/bindings_parser'
 require 'krane/label_selector'
@@ -69,7 +69,7 @@ template_paths = [template_dir] if template_paths.empty? && template_dir
 
 begin
   Krane::OptionsHelper.with_processed_template_paths(template_paths) do |paths|
-    runner = Krane::DeployTask.new(
+    runner = Krane::DeprecatedDeployTask.new(
       namespace: namespace,
       context: context,
       current_sha: ENV["REVISION"],
